Moved computer labs to my account
