+++ src/components/pages/account.js
@@ -45,10 +45,12 @@ const AccountSidebar = ({ user }) => (
         Your application status
       </AccountSidebarLink>
     )}
-    <AccountSidebarLink to="/account/card">Otter Card</AccountSidebarLink>
+
     {user._isStudent && (
       <AccountSidebarLink to="/account/laundry">Laundry</AccountSidebarLink>
     )}
+    <AccountSidebarLink to="/account/labs">Computer labs</AccountSidebarLink>
+    <AccountSidebarLink to="/account/card">Otter Card</AccountSidebarLink>
     <AccountSidebarLink to="/account/print">Printer paper</AccountSidebarLink>
     <AccountSidebarLink to="/account/emergency">
       Emergency alerts
