+++ src/style/typography.js
@@ -37,10 +37,6 @@ const type = {
     form: {
       margin: 0,
     },
-    'h2,h3,h4,h5': {
-      marginBottom: rhythm(1.2),
-      marginTop: 0,
-    },
   }),
 }
